[Silver II] Title: 나무 자르기, Time: 504 ms, Memory: 119296 KB -BaekjoonHub

+++ 백준/Silver/2805. 나무 자르기/나무 자르기.java
@@ -30,7 +30,7 @@ public class Main {
         int L = 0, R = 1000000000, ans = 0;
         while (L <= R) {
             int mid = (L + R) / 2;
-            if (determination((int)mid)) {
+            if (determination(mid)) {
                 ans = mid;
                 L = mid + 1;
             }
